Create `segments` module in `subspace-core-primitives` and move data structures around for better logical organization
